requirements.txt
Flask==2.3.3
flask-socketio==5.3.4
python-socketio==5.9.0
opencv-python==4.8.1.78
pyserial==3.5
numpy==1.24.3
python-engineio==4.7.1
Werkzeug==2.3.7
eventlet==0.33.3